Split container and start testing
