+++ src/app/containers/Footer/index.stories.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react'; // eslint-disable-line import/no-extraneous-dependencies
+import FooterContainer from './index';
+
+storiesOf('FooterContainer', module).add('default', () => <FooterContainer />);
