+++ src/app/containers/Footer/index.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import FooterContainer from './index';
+
+import { shouldMatchSnapshot } from '../../helpers/tests/testHelpers';
+
+describe(`FooterContainer`, () => {
+  shouldMatchSnapshot('should render correctly', <FooterContainer />);
+});
